Enable Isolated Projects for the Gradle build
Replaces the plugin's `afterEvaluate` + `rootProject.findProject(...)` wiring
with AGP component-API hooks and a dedicated resolvable configuration, so the
plugin no longer reads another project's state at configuration time.

- Split AGP-touching code into AndroidPreviewSupport so ComposePreviewPlugin
  stays loadable on classpaths without AGP (functional tests, Compose-MP-only
  consumers). Gradle decorates the plugin class at apply time and resolves
  referenced classes eagerly.
- Wire Android tasks via `pluginManager.withPlugin` + `finalizeDsl` +
  `onVariants`; sample's boot classpath now comes from
  `sdkComponents.bootClasspath` (drops the `rootProject.file("local.properties")`
  SDK lookup).
- Consume :renderer-android via a dedicated `composePreviewAndroidRenderer*`
  configuration whose attributes are copied from the sample's unit-test
  runtime classpath. Renderer class dirs are resolved from
  `project.rootDir` (IP-safe filesystem path), and cross-project build ordering
  uses a task-path `dependsOn`.
- `renderAllPreviews` picks between `historizePreviews` and `renderPreviews`
  via a Provider evaluated at graph-build time, so `historyEnabled.get()` is
  no longer read at plugin-apply time.
- Disable the Compose Multiplatform hot-reload sub-plugin
  (`org.jetbrains.compose.hot.reload.disable=true`), which was the last IP
  violator in the build.
