+++ gradle-plugin/src/main/kotlin/ee/schimke/composeai/plugin/ComposePreviewPlugin.kt
@@ -1,406 +1,42 @@
 package ee.schimke.composeai.plugin
 
-import com.android.build.api.dsl.CommonExtension
-import org.gradle.api.DefaultTask
 import org.gradle.api.Plugin
 import org.gradle.api.Project
-import org.gradle.api.attributes.Attribute
-import org.gradle.api.tasks.compile.JavaCompile
-import org.gradle.api.tasks.testing.Test
 
 class ComposePreviewPlugin : Plugin<Project> {
     override fun apply(project: Project) {
         val extension = project.extensions.create("composePreview", PreviewExtension::class.java)
 
-        project.afterEvaluate {
-            if (!extension.enabled.get()) return@afterEvaluate
+        // `pluginManager.withPlugin` replaces the old `project.afterEvaluate { ... }`
+        // block. `afterEvaluate` is discouraged under Gradle's Isolated Projects mode
+        // (and in general for plugin wiring); the plugin-manager hook fires as soon
+        // as the target plugin is applied, which is the right moment to wire up.
+        //
+        // The AGP-facing code (finalizeDsl / onVariants / cross-project dependency
+        // declaration) is isolated in [AndroidPreviewSupport]. Gradle decorates this
+        // plugin class at apply time and resolves all class references it sees in
+        // the plugin's bytecode — so keeping AGP types *out* of ComposePreviewPlugin
+        // is what lets the plugin load cleanly on non-Android projects (Compose
+        // Multiplatform consumers, functional tests, etc.). AGP classes only get
+        // loaded when `AndroidPreviewSupport.configure` actually runs.
+        var androidConfigured = false
+        val androidHandler: () -> Unit = {
+            if (!androidConfigured) {
+                androidConfigured = true
+                AndroidPreviewSupport.configure(project, extension)
+            }
+        }
+        project.pluginManager.withPlugin("com.android.application") { androidHandler() }
+        project.pluginManager.withPlugin("com.android.library") { androidHandler() }
 
-            val isAndroid = project.plugins.hasPlugin("com.android.application")
+        project.pluginManager.withPlugin("org.jetbrains.compose") {
+            if (androidConfigured) return@withPlugin
+            if (project.plugins.hasPlugin("com.android.application")
                 || project.plugins.hasPlugin("com.android.library")
-            val hasComposeDesktop = project.plugins.hasPlugin("org.jetbrains.compose")
-
-            if (isAndroid) {
-                enableAndroidResources(project)
-                registerAndroidTasks(project, extension)
-            } else if (hasComposeDesktop) {
-                registerDesktopTasks(project, extension)
-            } else {
-                project.logger.warn("compose-ai-tools: No Android or Compose Desktop target found")
+            ) {
+                return@withPlugin
             }
+            ComposePreviewTasks.registerDesktopTasks(project, extension)
         }
-    }
-
-    private fun enableAndroidResources(project: Project) {
-        val android = project.extensions.findByName("android") as? CommonExtension ?: return
-        android.testOptions.unitTests.isIncludeAndroidResources = true
-    }
-
-    private fun registerAndroidTasks(project: Project, extension: PreviewExtension) {
-        val variant = extension.variant.get()
-        val capVariant = variant.cap()
-        val previewOutputDir = project.layout.buildDirectory.dir("compose-previews")
-
-        val sourceClassDirs = project.files(
-            project.layout.buildDirectory.dir("tmp/kotlin-classes/$variant"),
-            project.layout.buildDirectory.dir("intermediates/javac/$variant/classes"),
-            project.layout.buildDirectory.dir("intermediates/built_in_kotlinc/$variant/compile${capVariant}Kotlin/classes"),
-        )
-
-        val dependencyConfigName = "${variant}RuntimeClasspath"
-
-        val discoverTask = registerDiscoverTask(project, sourceClassDirs, dependencyConfigName, previewOutputDir, extension) {
-            dependsOn("compile${capVariant}Kotlin")
-        }
-
-        val hasAndroidRenderer = project.rootProject.findProject(":renderer-android") != null
-
-        if (hasAndroidRenderer) {
-            val artifactType = Attribute.of("artifactType", String::class.java)
-            val rendererProject = project.rootProject.project(":renderer-android")
-            val rendererClassDirs = project.files(
-                rendererProject.layout.buildDirectory.dir("intermediates/built_in_kotlinc/$variant/compile${capVariant}Kotlin/classes"),
-                rendererProject.layout.buildDirectory.dir("tmp/kotlin-classes/$variant"),
-            )
-
-            // Build the classpath at configuration time using lazy file collections.
-            // Use debugUnitTestRuntimeClasspath with artifact view filtering to get
-            // extracted JARs from AARs (AGP registers the transforms).
-            val testConfig = project.configurations.findByName("${variant}UnitTestRuntimeClasspath")
-            val rendererConfig = rendererProject.configurations.findByName("${variant}RuntimeClasspath")
-
-            // The SDK stub android.jar is on the OUTER classpath so JUnit can introspect
-            // the test class (RobolectricRenderTest.kt references android.graphics.Bitmap,
-            // android.view.PixelCopy, etc. in method signatures). Without it, JUnit fails
-            // with `NoClassDefFoundError: android/graphics/Bitmap` during test discovery,
-            // before Robolectric's sandbox classloader is even created.
-            //
-            // Inside the sandbox, `ParameterizedRobolectricTestRunner` loads the test class
-            // through Robolectric's InstrumentingClassLoader, which delegates `android.*`
-            // resolution to its own `android-all` artifact (real framework classes, with
-            // shadows applied). The outer stub does NOT shadow the sandboxed PixelCopy.
-            val android = project.extensions.findByName("android") as? CommonExtension
-            val compileSdk = android?.compileSdk ?: 36
-            val sdkDir = findAndroidSdkDir(project)
-
-            val resolvedClasspath = project.files().apply {
-                if (testConfig != null) {
-                    from(testConfig.incoming.artifactView {
-                        attributes.attribute(artifactType, "jar")
-                    }.files)
-                    from(testConfig.incoming.artifactView {
-                        attributes.attribute(artifactType, "android-classes")
-                    }.files)
-                }
-                if (rendererConfig != null) {
-                    from(rendererConfig.incoming.artifactView {
-                        attributes.attribute(artifactType, "jar")
-                    }.files)
-                }
-                from(rendererClassDirs)
-                from(sourceClassDirs)
-                // android.jar for JUnit's outer-classloader test discovery (see above).
-                if (sdkDir != null) {
-                    val androidJar = java.io.File("$sdkDir/platforms/android-$compileSdk/android.jar")
-                    if (androidJar.exists()) from(androidJar)
-                }
-            }
-
-            // Copy JVM args from AGP's test task (at configuration time, safe for config cache)
-            val agpTestTask = project.tasks.findByName("test${capVariant}UnitTest") as? Test
-            val agpJvmArgs = agpTestTask?.jvmArgs ?: emptyList()
-
-            val manifestFile = previewOutputDir.map { it.file("previews.json").asFile.absolutePath }
-            val rendersDir = previewOutputDir.map { it.dir("renders").asFile.absolutePath }
-
-            val shardCount = resolveShardCount(project, extension, previewOutputDir.get().file("previews.json").asFile)
-            val shardsEnabled = shardCount > 1
-
-            // When sharded, generate N Java subclasses of RobolectricRenderTestBase, each with
-            // its own static @Parameters method that loads only that shard's slice of the manifest.
-            // Gradle distributes tests across forks at the class level, so a single parameterized
-            // class can't be split — we give it N classes. Each shard subclass inherits @Config
-            // and @GraphicsMode from the base, so every JVM's sandbox key matches and each fork
-            // reuses its own cached sandbox across all previews in its slice.
-            val shardSourcesDir = project.layout.buildDirectory.dir("generated/composeai/render-shards/java")
-            val shardClassesDir = project.layout.buildDirectory.dir("generated/composeai/render-shards/classes")
-
-            val generateShardsTask = if (shardsEnabled) {
-                project.tasks.register("generateRenderShards", GenerateRenderShardsTask::class.java) {
-                    group = "compose preview"
-                    description = "Generate $shardCount RobolectricRenderTest_Shard subclasses"
-                    shards.set(shardCount)
-                    outputDir.set(shardSourcesDir)
-                }
-            } else null
-
-            val compileShardsTask = if (generateShardsTask != null) {
-                project.tasks.register("compileRenderShards", JavaCompile::class.java) {
-                    group = "compose preview"
-                    description = "Compile generated shard test subclasses"
-                    source(generateShardsTask.map { it.outputDir.asFileTree })
-                    classpath = resolvedClasspath
-                    destinationDirectory.set(shardClassesDir)
-                    options.release.set(21)
-                    dependsOn(generateShardsTask)
-                    dependsOn(":renderer-android:compile${capVariant}Kotlin")
-                }
-            } else null
-
-            val renderTask = project.tasks.register("renderPreviews", Test::class.java) {
-                group = "compose preview"
-                description = "Render Android previews via Robolectric"
-                testClassesDirs = if (compileShardsTask != null) {
-                    rendererClassDirs + project.files(compileShardsTask.map { it.destinationDirectory })
-                } else {
-                    rendererClassDirs
-                }
-                classpath = if (compileShardsTask != null) {
-                    resolvedClasspath + project.files(compileShardsTask.map { it.destinationDirectory })
-                } else {
-                    resolvedClasspath
-                }
-                if (shardsEnabled) {
-                    include("**/RobolectricRenderTest_Shard*.class")
-                    maxParallelForks = shardCount
-                } else {
-                    include("**/RobolectricRenderTest.class")
-                }
-                useJUnit()
-
-                jvmArgs(agpJvmArgs)
-                jvmArgs(
-                    "--add-opens=java.base/java.lang=ALL-UNNAMED",
-                    "--add-opens=java.base/java.lang.reflect=ALL-UNNAMED",
-                    // Robolectric's `ShadowVMRuntime.getAddressOfDirectByteBuffer`
-                    // reflectively invokes `DirectByteBuffer.address()`; under JDK 17+
-                    // module rules this fails with IllegalAccessException without this
-                    // opens. Reached via `PathIterator` — triggered here by Wear Compose's
-                    // curved text renderer.
-                    "--add-opens=java.base/java.nio=ALL-UNNAMED",
-                )
-
-                // Configure Robolectric via system properties instead of @Config/@GraphicsMode
-                // annotations — annotations trigger eager android.app.Application resolution
-                // before the sandbox classloader exists.
-                // Graphics + looper modes plumbed via system properties rather than
-                // `@GraphicsMode` / `@LooperMode` annotations — historically the annotation
-                // path was suspected of eagerly resolving `android.app.Application` before
-                // the sandbox classloader was up. With NATIVE/PAUSED fixed in code via
-                // `@GraphicsMode(NATIVE)` on the test class we could drop these, but they
-                // are cheap and act as belt-and-suspenders.
-                systemProperty("robolectric.graphicsMode", "NATIVE")
-                systemProperty("robolectric.looperMode", "PAUSED")
-                // `robolectric.pixelCopyRenderMode=hardware` routes ShadowPixelCopy through
-                // `HardwareRenderingScreenshot` → `ImageReader` + `HardwareRenderer.syncAndDraw`,
-                // which is the only path that replays Compose's RenderNodes correctly. The
-                // legacy `robolectric.screenshot.hwrdr.native` flag (pre-4.10) is no longer
-                // read. The test class pins `@Config(sdk = [34])` because Robolectric 4.16.1
-                // stops shadowing `nativeCreatePlanes` above API 34, so capturing against
-                // SDK 35+ returns an Image with `planes[0] == null`.
-                systemProperty("robolectric.pixelCopyRenderMode", "hardware")
-
-                systemProperty("composeai.render.manifest", manifestFile.get())
-                systemProperty("composeai.render.outputDir", rendersDir.get())
-
-                dependsOn(discoverTask)
-                dependsOn(":renderer-android:compile${capVariant}Kotlin")
-                dependsOn("process${capVariant}Resources")
-                val configTaskName = "generate${capVariant}UnitTestConfig"
-                if (project.tasks.findByName(configTaskName) != null) {
-                    dependsOn(configTaskName)
-                }
-                if (compileShardsTask != null) {
-                    dependsOn(compileShardsTask)
-                }
-            }
-
-            registerRenderAllPreviews(project, extension, renderTask, previewOutputDir)
-        } else {
-            registerStubRenderTask(project, previewOutputDir, sourceClassDirs, dependencyConfigName, discoverTask, extension)
-        }
-    }
-
-    private fun registerDesktopTasks(project: Project, extension: PreviewExtension) {
-        val previewOutputDir = project.layout.buildDirectory.dir("compose-previews")
-
-        val sourceClassDirs = project.files(
-            project.layout.buildDirectory.dir("classes/kotlin/main"),
-            project.layout.buildDirectory.dir("classes/kotlin/jvm/main"),
-            project.layout.buildDirectory.dir("classes/kotlin/desktop/main"),
-        )
-
-        val dependencyConfigName = listOf("jvmRuntimeClasspath", "desktopRuntimeClasspath", "runtimeClasspath")
-            .firstOrNull { project.configurations.findByName(it) != null }
-            ?: "runtimeClasspath"
-
-        val discoverTask = registerDiscoverTask(project, sourceClassDirs, dependencyConfigName, previewOutputDir, extension) {
-            for (name in listOf("compileKotlinJvm", "compileKotlinDesktop", "compileKotlin")) {
-                if (project.tasks.findByName(name) != null) {
-                    dependsOn(name)
-                    break
-                }
-            }
-        }
-
-        val rendererConfigName = "composePreviewRenderer"
-        val rendererConfig = project.configurations.maybeCreate(rendererConfigName)
-        rendererConfig.isCanBeResolved = true
-        rendererConfig.isCanBeConsumed = false
-
-        val hasDesktopRenderer = try {
-            project.dependencies.add(rendererConfigName, project.dependencies.project(mapOf("path" to ":renderer-desktop")))
-            true
-        } catch (_: Exception) {
-            false
-        }
-
-        if (hasDesktopRenderer) {
-            val renderTask = project.tasks.register("renderPreviews", RenderPreviewsTask::class.java) {
-                previewsJson.set(previewOutputDir.map { it.file("previews.json") })
-                outputDir.set(previewOutputDir.map { it.dir("renders") })
-                renderBackend.set("desktop")
-                useComposeRenderer.set(true)
-                renderClasspath.from(sourceClassDirs)
-                project.configurations.findByName(dependencyConfigName)?.let { renderClasspath.from(it) }
-                renderClasspath.from(rendererConfig)
-                group = "compose preview"
-                description = "Render all previews to PNG"
-                dependsOn(discoverTask)
-            }
-            registerRenderAllPreviews(project, extension, renderTask, previewOutputDir)
-        } else {
-            registerStubRenderTask(project, previewOutputDir, sourceClassDirs, dependencyConfigName, discoverTask, extension)
-        }
-    }
-
-    private fun registerDiscoverTask(
-        project: Project,
-        sourceClassDirs: org.gradle.api.file.FileCollection,
-        dependencyConfigName: String,
-        previewOutputDir: org.gradle.api.provider.Provider<org.gradle.api.file.Directory>,
-        extension: PreviewExtension,
-        configureDeps: DiscoverPreviewsTask.() -> Unit,
-    ): org.gradle.api.tasks.TaskProvider<DiscoverPreviewsTask> {
-        val artifactType = Attribute.of("artifactType", String::class.java)
-
-        return project.tasks.register("discoverPreviews", DiscoverPreviewsTask::class.java) {
-            classDirs.from(sourceClassDirs)
-            project.configurations.findByName(dependencyConfigName)?.let { config ->
-                // For Android projects, dependencies resolve as AARs. Use artifact view
-                // filtering to request the extracted classes.jar (AGP registers the
-                // transform). Desktop/JVM projects already return JARs so this is a no-op.
-                dependencyJars.from(config.incoming.artifactView {
-                    attributes.attribute(artifactType, "jar")
-                }.files)
-                dependencyJars.from(config.incoming.artifactView {
-                    attributes.attribute(artifactType, "android-classes")
-                }.files)
-            }
-            moduleName.set(project.name)
-            variantName.set(extension.variant)
-            outputFile.set(previewOutputDir.map { it.file("previews.json") })
-            group = "compose preview"
-            description = "Discover @Preview annotations in compiled classes"
-            configureDeps()
-        }
-    }
-
-    private fun registerStubRenderTask(
-        project: Project,
-        previewOutputDir: org.gradle.api.provider.Provider<org.gradle.api.file.Directory>,
-        sourceClassDirs: org.gradle.api.file.FileCollection,
-        dependencyConfigName: String,
-        discoverTask: org.gradle.api.tasks.TaskProvider<DiscoverPreviewsTask>,
-        extension: PreviewExtension,
-    ) {
-        val renderTask = project.tasks.register("renderPreviews", RenderPreviewsTask::class.java) {
-            previewsJson.set(previewOutputDir.map { it.file("previews.json") })
-            outputDir.set(previewOutputDir.map { it.dir("renders") })
-            renderBackend.set("stub")
-            useComposeRenderer.set(false)
-            renderClasspath.from(sourceClassDirs)
-            project.configurations.findByName(dependencyConfigName)?.let { renderClasspath.from(it) }
-            group = "compose preview"
-            description = "Render all previews to PNG (stub)"
-            dependsOn(discoverTask)
-        }
-        registerRenderAllPreviews(project, extension, renderTask, previewOutputDir)
-    }
-
-    /**
-     * Registers `renderAllPreviews` as the user-facing entry point. When
-     * `composePreview.historyEnabled = true`, inserts a `historizePreviews`
-     * task between the render task and the aggregate so every run archives
-     * changed PNGs into the configured history directory.
-     */
-    private fun registerRenderAllPreviews(
-        project: Project,
-        extension: PreviewExtension,
-        renderTask: org.gradle.api.tasks.TaskProvider<*>,
-        previewOutputDir: org.gradle.api.provider.Provider<org.gradle.api.file.Directory>,
-    ) {
-        val finalTask: org.gradle.api.tasks.TaskProvider<*> = if (extension.historyEnabled.get()) {
-            // Default history dir: <project>/.compose-preview-history — outside `build/`
-            // so snapshots survive `./gradlew clean`. Users can override via extension.
-            val defaultHistoryDir = project.layout.projectDirectory.dir(".compose-preview-history")
-            val resolvedHistoryDir = extension.historyDir.orElse(defaultHistoryDir)
-
-            project.tasks.register("historizePreviews", HistorizePreviewsTask::class.java) {
-                previewsJson.set(previewOutputDir.map { it.file("previews.json") })
-                rendersDir.set(previewOutputDir.map { it.dir("renders") })
-                historyDir.set(resolvedHistoryDir)
-                group = "compose preview"
-                description = "Archive changed previews into the local history directory"
-                dependsOn(renderTask)
-            }
-        } else {
-            renderTask
-        }
-
-        project.tasks.register("renderAllPreviews", DefaultTask::class.java) {
-            group = "compose preview"
-            dependsOn(finalTask)
-        }
-    }
-
-    private fun findAndroidSdkDir(project: Project): String? {
-        return project.providers.environmentVariable("ANDROID_HOME").orNull
-            ?: project.providers.environmentVariable("ANDROID_SDK_ROOT").orNull
-            ?: project.rootProject.file("local.properties").let { f ->
-                if (f.exists()) f.readLines()
-                    .find { it.startsWith("sdk.dir=") }
-                    ?.substringAfter("sdk.dir=")
-                else null
-            }
-    }
-
-    private fun String.cap(): String =
-        replaceFirstChar { if (it.isLowerCase()) it.titlecase() else it.toString() }
-
-    /**
-     * Resolves the effective shard count from [PreviewExtension.shards]:
-     *
-     *  - `≥1`: use the value as-is.
-     *  - `0` (auto): read [previewsJson] if it exists from a previous discover run and
-     *    hand the count to [ShardTuning.autoShards]. If the file is missing (very first
-     *    build), fall back to 1 — the next run will have better data and can
-     *    pick a higher count then.
-     */
-    private fun resolveShardCount(
-        project: Project,
-        extension: PreviewExtension,
-        previewsJson: java.io.File,
-    ): Int {
-        val requested = extension.shards.get()
-        if (requested > 0) return requested
-        if (!previewsJson.exists()) {
-            project.logger.info("compose-ai-tools: shards=auto but previews.json missing; defaulting to 1 for this run")
-            return 1
-        }
-        // Simple count of `"id"` entries — cheap and avoids pulling kotlinx.serialization into the plugin classpath.
-        val previewCount = Regex("\"id\"\\s*:").findAll(previewsJson.readText()).count()
-        val resolved = ShardTuning.autoShards(previewCount)
-        project.logger.lifecycle("compose-ai-tools: shards=auto → $resolved (previewCount=$previewCount, cores=${Runtime.getRuntime().availableProcessors()})")
-        return resolved
     }
 }
