+++ gradle.properties
@@ -8,3 +8,10 @@ android.useAndroidX=true
 
 # Workaround for Gradle 9 and Kotlin plugin compatibility issue, should be removed when fixed and won't be supported in Gradle 10
 org.gradle.configuration-cache.unsafe.ignore.unsupported-build-events-listeners=true
+
+# Compose Multiplatform auto-applies a hot-reload sub-plugin that violates
+# Isolated Projects (accesses :renderer-desktop / :sample-cmp's layout from :).
+# We don't use hot reload, so disable it.
+org.jetbrains.compose.hot.reload.disable=true
+
+org.gradle.unsafe.isolated-projects=true
